lab-sts/requirements.txt
boto3==1.26.116
botocore==1.29.116
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
